Use size_t in CycleTabs

+++ src/tsframe.h
@@ -860,8 +860,8 @@ struct TSFrame : wxFrame {
         }
     }
 
-    void CycleTabs(int offset = 1) {
-        auto numtabs = static_cast<int>(notebook->GetPageCount());
+    void CycleTabs(size_t offset = 1) {
+        auto numtabs = notebook->GetPageCount();
         offset = ((offset >= 0) ? 1 : numtabs - 1);  // normalize to non-negative wrt modulo
         notebook->SetSelection((notebook->GetSelection() + offset) % numtabs);
     }
